monitoring/python-web-2/app/requirements.txt
Flask==2.3.3
pymongo==4.5.0
prometheus-client==0.17.1
python-dotenv==1.0.0
requests==2.31.0